refactor: MenuGroup domain 리팩토링
